Copyright notice included.
I added a copyright notice to my README.md file to let potential infringers know that this repository and its contents will not be copied, reproduced, modified, or distributed in any way without permission. In addition, the code will not be used by unauthorized users as a way of generating commercial profit.